Adding HTML documentation to repository.
This commit modifies the documentation make files to put html builds
directly in the doc/ directory instead of doc/build so they aren't
ignored by .gitignore.

The html build was added to the repository.  I'm hoping that by
keeping it current before each commit, I can have a link to the most
recent development docs on the webpage that will automatically stay
updated with the code.