Overhaul of the discrete trajectory optimization tools.
This is a major update to the discrete trajectory optimization tools
in trep.  In particular, DOptimzer has been almost completely
reorganized.  The new design makes it easier to use the various
subparts outside of the built in optimization method so that users can
play around easier.

The new version is still unoptimized and missing the monitoring code.