Fixed DOptimizer.check_ddcost() method.
I believe this finally fixes the check_ddcost() method.  The missing
link was that the descent direction has to be projected into the
tangent trajectory space at the perturbed trajectory.  The check
pend-on-cart-optimization.py is giving much more reasonable results.
The approximations still aren't as clean as you see with other
derivatives throughout trep, but I think that is to be expected
because there is so much more room for numerical errors to creep in.