Automatically updating version information.
trep version information is now contained in src/__version__.py.  If
git is present, setup.py will always update this file with a version
string created by git-describe, otherwise it will leave the version
alone (or create a new file with version "<unknown>" if git is
unavailable and __version__.py does not exist.)  The version
information is automatically passed to the documentation build as
well.