Enhancements to DSystem and DOptimizer.
Created DSystem.calc_feedback_controller() to calculate a feedback
controller about a (not necessarily valid) trajectory.
DOptimizer.calc_projection() is now just a wrapper around this.

Moved DOptimizer.project() to DSystem.project().  The projection is
really useful and having it here makes it easier to use since you
don't have to have a optimizer.  The method lets you specify a
feedback law, or else tries to construct one about the given
trajectory, which should work fine as long as it's almost a
trajectory.

Modified DSystem.calc_linearization() to call self.set() at every
timestep instead of simulating from the intial condition with
self.step().  This should be more numerically stable and should be
able to construct a linearization about almost-trajectories.

Added new method DSystem.import_trajectory(dsys_a, X, U) that maps a
trajectory X,U for dsys_a to a new trajectory for this system.  States
and inputs in dsys_a that are not in self are dropped.  States and
inputs that are in self but not dsys_a are left as zeros.