Added solution hints to MidpointVI and DSystem.
This commit adds new parameters to MidpointVI.step(), DSystem.set(),
and DSystem.step() that optionally let the caller provide a hint for
what the next state will be.  When working along a known trajectory,
this can reduce the amount of time spent in the DEL root solver.

An (untracked) puppet trajectory optimization saw a 30% speed increase
from this change.  The pend-on-cart-optimization example shows
insignicant improvement.