Reintroducing DOptimizerMonitor and improvements to DOptimizer.
The optimization monitor was removed in the recent overhaul of
DOptimizer.  This commit reintroduces it.  There are two monitors now.

The base class, DOptimizerMonitor does nothing, so it can be used as a
silent monitor.

The derived DOptimizerDefaultMonitor prints progress information to
stdout, replacing the print statements that used to be throughout
DOptimizer.  It also saves the cost and cost derivative history.  As
the name implies, this is the default monitor used by DOptimizer.

A lot of old unused attributes were cleared out of DOptimizer.

Also, the descent method selection and fallback method selection were
factored out into separate methods of DOptimizer so that a user can
derive a new class and override them if they want to change the
behavior.