Created external extension module for piston force.
This commit separates the piston force from the radial engine example.
The piston is implemented as a separate extension module.  The old
example is still in trep, but I'll probably remove it in the future.
Right now its nice as a benchmark to see if there is any speed
difference between being built into trep and being built externally.