Adding optimization to TapeMeasure derivatives
This commit optimizes the first and second derivatives of the
TapeMeasure length by building a table of segments that depend on each
configuration and only looping over those segments.  The optimization
is based on the fact that a segment depends on a configuration
variable if only one (not both) of the segment frames depends on the
config.  This shows a significant reduction of work in the hand
models.