Modify MidpointVI step function to handle singular matrices correctly.
When the derviative of the DEL equation was singular, LU_decomp sets a
ValueError exception instead of a ConvergenceException.  This is
correct behavior, so a try/except block was added to MidpointVI to
catch this exception and raise ConvergenceError so that calling
functions only have to watch for one exception instead of two.