Fixing #34
Small issues existed in both trep/forces/spatialwrench.py and
trep/forces/bodywrench.py when accessing the current values of the wrench. These
issues did not exist in trep/forces/hybridwrench.py, but I used the opportunity
to clean up some trailing whitespace.

While debugging this, I also discovered another small issue. When passing a list
of inputs to the VisualItem constructor, the user was previously expected to
pass a list that was one time index shorter than that used for the time and
configuration lists. This makes some sense because for a given time window
from (k) to (k+1), the solved VI will know (q(k), p(k)) and (q(k+1), p(k+1)),
and it will also know u(k) but not u(k+1). However, this was causing issues in
setTime() when the interpolation function for u was then asked to interpolate
outside of the range of times originally passed in. We could fix this with some
simple try-except blocks, or always handle the u interpolation slightly
differently than the q interpolation. Instead, I chose to require an equal
length u array. The user can always zero-pad. Would be better to more robustly
handle this in the future, but for now, this is fine.